# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "StartupCRM API",
    "description": "API para gestión de CRM - Autenticación, Emails, Contactos",
    "version": "1.0.0",
    "contact": {
      "name": "StartupCRM Support",
      "email": "support@startupcrm.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/api/email/send": {
      "post": {
        "summary": "Enviar email",
        "description": "Envía un email usando configuración SMTP. **Requiere autenticación**",
        "tags": ["Email"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["to", "subject"],
                "properties": {
                  "to": {
                    "type": "string",
                    "description": "Destinatario",
                    "example": "example@gmail.com"
                  },
                  "subject": {
                    "type": "string",
                    "description": "Asunto del email",
                    "example": "Prueba HTML4"
                  },
                  "html": {
                    "type": "string",
                    "description": "Contenido HTML",
                    "example": "<p>Hola desde CRM</p>"
                  },
                  "text": {
                    "type": "string",
                    "description": "Contenido texto plano",
                    "example": "Hola desde CRM"
                  },
                  "contactId": {
                    "type": "string",
                    "description": "ID de contacto"
                  },
                  "templateId": {
                    "type": "string",
                    "description": "ID de plantilla",
                    "example": "welcome-01"
                  },
                  "metadata": {
                    "type": "object",
                    "description": "Metadatos adicionales",
                    "example": {
                      "campaign": "onboarding"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email enviado exitosamente",
            "content": {
              "application/json": {
                "example": {
                  "status": "sent",
                  "contactId": "uuid",
                  "contactCreated": true,
                  "messageId": "<123@example.com>",
                  "message": "Email enviado correctamente"
                }
              }
            }
          },
          "400": {
            "description": "Error en validación"
          },
          "401": {
            "description": "No autorizado - requiere autenticación"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
...
    "/api/email/sync-status": {
      "get": {
        "summary": "Estado de sincronización",
        "description": "Retorna el estado actual de la sincronización de emails. **Requiere autenticación**",
        "tags": ["Email"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Estado de sync",
            "content": {
              "application/json": {
                "example": {
                  "lastProcessedMessageId": "abc123@email.com",
                  "lastSyncAt": "2026-04-05T01:20:00Z",
                  "totalEmailsEntrantes": 150,
                  "recentEmails": []
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
...
    "/api/whatsapp": {
      "get": {
        "summary": "Listar contactos WhatsApp",
        "description": "Retorna lista de contactos con conversaciones de WhatsApp. **Requiere autenticación**",
        "tags": ["WhatsApp"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": { "type": "number", "default": 1 },
            "description": "Número de página"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "number", "default": 20, "enum": [10, 20, 30, 50, 100] },
            "description": "Registros por página"
          },
          {
            "name": "search",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Buscar por nombre o teléfono"
          },
          {
            "name": "unreadOnly",
            "in": "query",
            "schema": { "type": "boolean", "default": false },
            "description": "Solo no leídos"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de contactos WhatsApp",
            "content": {
              "application/json": {
                "example": {
                  "data": [
                    {
                      "contactId": "uuid",
                      "name": "Juan Perez",
                      "phone": "+59176543210",
                      "lastMessage": "Hola mundo",
                      "unreadCount": 2,
                      "lastMessageAt": "2026-04-07T10:30:00Z"
                    }
                  ],
                  "pagination": { "page": 1, "limit": 20, "total": 50, "totalPages": 3 }
                }
              }
            }
          },
          "401": { "description": "No autorizado" }
        }
      },
      "post": {
        "summary": "Sincronizar mensajes WhatsApp",
        "description": "Sincroniza mensajes entrantes desde WPPConnect. **Requiere autenticación**",
        "tags": ["WhatsApp"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Sincronización completada",
            "content": {
              "application/json": {
                "example": {
                  "status": "ok",
                  "newMessages": 5,
                  "contactsUpdated": 2
                }
              }
            }
          },
          "401": { "description": "No autorizado" },
          "500": { "description": "Error del servidor" }
        }
      }
    },
    "/api/whatsapp/{contactId}": {
      "get": {
        "summary": "Obtener mensajes de contacto",
        "description": "Retorna mensajes de WhatsApp para un contacto específico. **Requiere autenticación**",
        "tags": ["WhatsApp"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "contactId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "description": "ID del contacto"
          },
          {
            "name": "page",
            "in": "query",
            "schema": { "type": "number", "default": 1 },
            "description": "Número de página"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "number", "default": 20 },
            "description": "Registros por página"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de mensajes",
            "content": {
              "application/json": {
                "example": {
                  "data": [
                    {
                      "id": "uuid",
                      "contenido": "Hola, tienes información?",
                      "direccion": "entrante",
                      "fecha": "2026-04-07T10:30:00Z",
                      "leido": false
                    }
                  ],
                  "pagination": { "page": 1, "limit": 20, "total": 100, "totalPages": 5 }
                }
              }
            }
          },
          "401": { "description": "No autorizado" },
          "404": { "description": "Contacto no encontrado" }
        }
      }
    },
    "/api/whatsapp/send": {
      "post": {
        "summary": "Enviar mensaje WhatsApp",
        "description": "Envía un mensaje de texto por WhatsApp. **Requiere autenticación**",
        "tags": ["WhatsApp"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "contactId": {
                    "type": "string",
                    "description": "ID del contacto (opcional si se usa to)",
                    "example": "uuid-del-contacto"
                  },
                  "to": {
                    "type": "string",
                    "description": "Número de teléfono con código de país (opcional si se usa contactId)",
                    "example": "+59176543210"
                  },
                  "text": {
                    "type": "string",
                    "description": "Contenido del mensaje",
                    "example": "Hola, te contactamos desde el CRM"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Mensaje enviado",
            "content": {
              "application/json": {
                "example": {
                  "status": "sent",
                  "messageId": "wamid.abc123",
                  "contactId": "uuid",
                  "message": "Mensaje de WhatsApp enviado correctamente"
                }
              }
            }
          },
          "400": { "description": "Error de validación" },
          "401": { "description": "No autorizado" },
          "500": { "description": "Error del servidor" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Better Auth session token (cookie: better-auth.session_token)"
      }
    },
...
  },
...
}